+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Verb Changelog
-## **master** (MELPA)
+## **main** (MELPA)
 - Fixed error when receiving JSON responses that include `"t"` as a key.
 - Allow underscores (`_`) in HTTP header names.
 - Fixed <kbd>C-u C-c C-r C-r</kbd> and <kbd>C-u C-c C-r C-s</kbd> not displaying the response buffer when a response is received.